requirements.txt
python-telegram-bot[job-queue]
openai
gspread
oauth2client
python-dotenv